+++ SqlSizer-MSSQL/Public/Find-UnreachableTables.ps1
@@ -4,26 +4,26 @@ function Find-UnreachableTables
     param
     (
         [Parameter(Mandatory = $true)]
-        [Query[]]$Queries,
+        [Query2[]]$Queries,
 
         [Parameter(Mandatory = $false)]
-        [ColorMap]$ColorMap = $null,
+        [TraversalConfiguration]$TraversalConfig = $null,
 
         [Parameter(Mandatory = $true)]
         [DatabaseInfo]$DatabaseInfo
     )
 
     $reachableTables = New-Object 'System.Collections.Generic.HashSet[string]'
-    $processedTableColors = New-Object 'System.Collections.Generic.HashSet[string]'
-    $processingQueue = New-Object System.Collections.Generic.Queue"[TableInfo2WithColor]"
+    $processedTableStates = New-Object 'System.Collections.Generic.HashSet[string]'
+    $processingQueue = New-Object System.Collections.Generic.Queue"[TableInfo2WithState]"
 
     # Add all tables to processing
     foreach ($query in $Queries)
     {
-        $item = New-Object TableInfo2WithColor
+        $item = New-Object TableInfo2WithState
         $item.SchemaName = $query.Schema
         $item.TableName = $query.Table
-        $item.Color = $query.Color
+        $item.State = $query.State
 
         $null = $processingQueue.Enqueue($item)
     }
@@ -36,88 +36,93 @@ function Find-UnreachableTables
         }
 
         $item = $processingQueue.Dequeue()
-        $key = $item.SchemaName + "." + $item.TableName + "." + $item.Color
+        $key = $item.SchemaName + "." + $item.TableName + "." + $item.State
 
-        if ($processedTableColors.Contains($key))
+        if ($processedTableStates.Contains($key))
         {
             continue
         }
 
         $table = $DatabaseInfo.Tables.Where(({ ($_.TableName -eq $item.TableName) -and ($_.SchemaName -eq $item.SchemaName) }))[0]
 
-        if ($item.Color -eq [Color]::Yellow)
+        if ($item.State -eq [TraversalState]::Pending)
         {
-            $newItem = New-Object TableInfo2WithColor
+            $newItem = New-Object TableInfo2WithState
             $newItem.SchemaName = $item.SchemaName
             $newItem.TableName = $item.TableName
-            $newItem.Color = [Color]::Green
+            $newItem.State = [TraversalState]::Include
             $null = $processingQueue.Enqueue($newItem)
 
-            $newItem = New-Object TableInfo2WithColor
+            $newItem = New-Object TableInfo2WithState
             $newItem.SchemaName = $item.SchemaName
             $newItem.TableName = $item.TableName
-            $newItem.Color = [Color]::Red
+            $newItem.State = [TraversalState]::Exclude
             $null = $processingQueue.Enqueue($newItem)
         }
 
-        if (($item.Color -eq [Color]::Red) -or ($item.Color -eq [Color]::Purple))
+        if ($item.State -eq [TraversalState]::Exclude)
         {
             foreach ($fk in $table.ForeignKeys)
             {
-                $newItem = New-Object TableInfo2WithColor
+                $newItem = New-Object TableInfo2WithState
                 $newItem.SchemaName = $fk.Schema
                 $newItem.TableName = $fk.Table
 
-                $newColor = [Color]::Red
-                if ($null -ne $ColorMap)
+                $newState = [TraversalState]::Exclude
+                if ($null -ne $TraversalConfig)
                 {
-                    $items = $ColorMap.Items | Where-Object { ($_.SchemaName -eq $fk.Schema) -and ($_.TableName -eq $fk.Table) }
-                    $items = $items | Where-Object { ($null -eq $_.Condition) -or ((($_.Condition.SourceSchemaName -eq $fk.FkSchema) -or ("" -eq $_.Condition.SourceSchemaName)) -and (($_.Condition.SourceTableName -eq $fk.FkTable) -or ("" -eq $_.Condition.SourceTableName))) }
-                    if (($null -ne $items) -and ($null -ne $items.ForcedColor))
+                    $rule = $TraversalConfig.GetItemForTable($fk.Schema, $fk.Table)
+                    if ($null -eq $rule)
                     {
-                        $newColor = [int]$items.ForcedColor.Color
+                        $rules = $TraversalConfig.Rules | Where-Object { ($_.SchemaName -eq $fk.Schema) -and ($_.TableName -eq $fk.Table) }
+                        $rules = $rules | Where-Object { ($null -eq $_.Constraints) -or ((($_.Constraints.SourceSchemaName -eq $fk.FkSchema) -or ("" -eq $_.Constraints.SourceSchemaName)) -and (($_.Constraints.SourceTableName -eq $fk.FkTable) -or ("" -eq $_.Constraints.SourceTableName))) }
+                        $rule = $rules | Select-Object -First 1
+                    }
+                    if (($null -ne $rule) -and ($null -ne $rule.StateOverride))
+                    {
+                        $newState = $rule.StateOverride.State
                     }
                 }
 
-                $newItem.Color = $newColor
+                $newItem.State = $newState
                 $null = $processingQueue.Enqueue($newItem)
             }
         }
 
-        if (($item.Color -eq [Color]::Green) -or ($item.Color -eq [Color]::Blue) -or ($item.Color -eq [Color]::Purple))
+        if (($item.State -eq [TraversalState]::Include) -or ($item.State -eq [TraversalState]::InboundOnly))
         {
             foreach ($referencedByTable in $table.IsReferencedBy)
             {
                 $fks = $referencedByTable.ForeignKeys | Where-Object { ($_.Schema -eq $item.SchemaName) -and ($_.Table -eq $item.TableName) }
                 foreach ($fk in $fks)
                 {
-                    $newItem = New-Object TableInfo2WithColor
+                    $newItem = New-Object TableInfo2WithState
                     $newItem.SchemaName = $fk.FkSchema
                     $newItem.TableName = $fk.FkTable
 
-                    if ($item.Color -eq [Color]::Green)
+                    if ($item.State -eq [TraversalState]::Include)
                     {
-                        $newItem.Color = [Color]::Yellow
+                        $newItem.State = [TraversalState]::Pending
                     }
-                    if ($item.Color -eq [Color]::Blue)
+                    if ($item.State -eq [TraversalState]::InboundOnly)
                     {
-                        $newItem.Color = [Color]::Blue
-                    }
-
-                    if ($item.Color -eq [int][Color]::Purple)
-                    {
-                        $newItem.Color = [int][Color]::Red
+                        $newItem.State = [TraversalState]::InboundOnly
                     }
 
-                    # forced color from color map
-                    if ($null -ne $ColorMap)
+                    # forced state from traversal config
+                    if ($null -ne $TraversalConfig)
                     {
-                        $items = $ColorMap.Items | Where-Object { ($_.SchemaName -eq $fk.FkSchema) -and ($_.TableName -eq $fk.FkTable) }
-                        $items = $items | Where-Object { ($null -eq $_.Condition) -or ((($_.Condition.SourceSchemaName -eq $fk.Schema) -or ("" -eq $_.Condition.SourceSchemaName)) -and (($_.Condition.SourceTableName -eq $fk.Table) -or ("" -eq $_.Condition.SourceTableName))) }
-
-                        if (($null -ne $items) -and ($null -ne $items.ForcedColor))
+                        $rule = $TraversalConfig.GetItemForTable($fk.FkSchema, $fk.FkTable)
+                        if ($null -eq $rule)
                         {
-                            $newItem.Color = [int]$items.ForcedColor.Color
+                            $rules = $TraversalConfig.Rules | Where-Object { ($_.SchemaName -eq $fk.FkSchema) -and ($_.TableName -eq $fk.FkTable) }
+                            $rules = $rules | Where-Object { ($null -eq $_.Constraints) -or ((($_.Constraints.SourceSchemaName -eq $fk.Schema) -or ("" -eq $_.Constraints.SourceSchemaName)) -and (($_.Constraints.SourceTableName -eq $fk.Table) -or ("" -eq $_.Constraints.SourceTableName))) }
+                            $rule = $rules | Select-Object -First 1
+                        }
+
+                        if (($null -ne $rule) -and ($null -ne $rule.StateOverride))
+                        {
+                            $newItem.State = $rule.StateOverride.State
                         }
                     }
 
@@ -127,7 +132,7 @@ function Find-UnreachableTables
         }
 
         $null = $reachableTables.Add($item.SchemaName + "." + $item.TableName)
-        $null = $processedTableColors.Add($key)
+        $null = $processedTableStates.Add($key)
     }
 
     $toReturn = @()